Fixed problem with rewrite_seed_with_rf.pl that causes it to fail for certain zero basepair models. With such models the SEED can be modified (or 'doctored') in cmbuild by moving 'inserts' into 'match' positions in a manner analogous to what HMMER does in hmmbuild (this behavior was implemented to make the HMM filters work better for zero basepair CMs, which are effectively HMMs).
The fix is to abandon the use of the 'cmbuild -O' output alignment to
map the RF annotation to the original SEED. Now,
rewrite_seed_with_rf.pl and rfsearch.pl use an output alignment from
'cmalign --mapali SEED CM <small.fa>' with the sequences from
<small.fa> removed, to map the RF to the original SEED. cmalign does
not doctor the SEED alignment, so the above problem is avoided.

+++ trunk/Rfam/Scripts/jiffies/rewrite_seed_with_rf.pl
@@ -13,17 +13,18 @@
 # 1. Verifies SEED is younger than CM and CM is younger than 
 #    'TBLOUT', 'SCORES', etc.
 # 2. Reads in a SEED and a calibrated CM built from that SEED
-#    and verifies the CM came from the SEED.     
-# 3. Uses 'cmbuild' to create a new SEED, with -O option.
-# 4. Verifies new CM is identical to original CM.
-# 5. Adds RF annotation from new SEED to original SEED to create
+#    and verifies the CM came from the SEED using 'cmalign
+#    --mapali SEED CM c.fa', where c.fa is the consensus 
+#    sequence for the CM. This gives us a new version of the SEED
+#    plus the single consensus sequence, with RF annotation.
+# 3. Adds RF annotation from cmalign output to original SEED to create
 #    a new SEED file, and makes all inserts lowercase and 
 #    consensus residues uppercase, to make even newer SEED.
-# 6. Builds even newer CM from even newer SEED and verifies 
+# 4. Builds even newer CM from even newer SEED and verifies 
 #    even newer CM is same as new CM (and also same as original CM).
-# 7. Adds E-value stats from original CM to even newer CM.
-# 8. Validates that we can use --mapali with new CM and new SEED
-# 9. Uses 'touch' to "update" 'TBLOUT', 'SCORES', 'REVTBLOUT', 
+# 5. Adds E-value stats from original CM to even newer CM.
+# 6. Validates that we can use --mapali with new CM and new SEED
+# 7. Uses 'touch' to "update" 'TBLOUT', 'SCORES', 'REVTBLOUT', 
 #    'outlist', 'species', 'taxinfo', 'revspecies', 'outlist.pdf',
 #    etc.
 
@@ -50,9 +51,9 @@ use Bio::Easel::MSA;
 # than those in tier 3.
 my @reqdFiles_t1 = ("SEED");
 my @reqdFiles_t2 = ("CM");
-my @reqdFiles_t3 = ("DESC", "outlist", "outlist.pdf", "revoutlist",
-                    "revspecies", "REVTBLOUT", "SCORES", "species",
+my @reqdFiles_t3 = ("DESC", "outlist", "outlist.pdf", "REVTBLOUT", "SCORES", "species",
                     "species.pdf", "taxinfo", "TBLOUT");
+my @nonreqdFiles = ("revoutlist", "revspecies"); # these will only exist if there were >0 reverse hits
 
 foreach my $file (@reqdFiles_t1, @reqdFiles_t2, @reqdFiles_t3) { 
   if(! -e $file) { die "ERROR, required file $file does not exist"; }
@@ -105,21 +106,30 @@ if(scalar(@comA)    != 2) { die "ERROR creating COM array"; }
 
 
 ####################################################################################################
-# 1. Verifies SEED is younger than CM and CM is younger than 
-#    'TBLOUT', 'SCORES', etc.
+# 1. Verifies CM is younger than 'TBLOUT', 'SCORES', etc.
+#    We used to verify SEED was younger than CM, but that's actually
+#    not strictly necessary, a more important test is whether the CM
+#    was built from the SEED, which is a test we DO do below. So 
+#    as of now we don't require SEED be older than CM here (there
+#    were about 25 families in the rethresholded dir for which 
+#    SEED was younger than CM...).
+
 my ($file1, $file2);
 # verify all required files exist
 foreach $file1 (@reqdFiles_t1, @reqdFiles_t2, @reqdFiles_t3) { 
   if(! -s $file1) { die "ERROR required file $file1 does not exist or is empty!"; }
 }
+
+# IF WE WANT TO MAKE SURE SEED IS YOUNGER THAN CM, UNCOMMENT THIS BLOCK
 # verify files in tier 1 are older than those in tier 2
-foreach $file1 (@reqdFiles_t1) { 
-  foreach $file2 (@reqdFiles_t2) { 
-    if(! Bio::Rfam::Utils::youngerThan($file2, $file1)) { 
-      die "ERROR file in tier 1: $file1 not older than file in tier 2: $file2\n"; 
-    }
-  }
-}
+#foreach $file1 (@reqdFiles_t1) { 
+#  foreach $file2 (@reqdFiles_t2) { 
+#    if(! Bio::Rfam::Utils::youngerThan($file2, $file1)) { 
+#      die "ERROR file in tier 1: $file1 not older than file in tier 2: $file2\n"; 
+#    }
+#  }
+#}
+
 # verify files in tier 2 are older than those in tier 3
 foreach $file1 (@reqdFiles_t2) { 
   foreach $file2 (@reqdFiles_t3) { 
@@ -129,17 +139,20 @@ foreach $file1 (@reqdFiles_t2) {
   }
 }
 
-####################################################################################################
+####################################################
 # 2. Reads in a SEED and a calibrated CM built from that SEED
-#    and verifies the CM came from the SEED.     
+#    and verifies the CM came from the SEED using 'cmalign
+#    --mapali SEED CM c.fa', where c.fa is the consensus 
+#    sequence for the CM. This gives us a new version of the SEED
+#    plus the single consensus sequence, with RF annotation.
 #
-# make sure CM file was built from SEED file.
-#
-# 1. verify a --mapali will work, this will make sure the checksum of the input alignment 
-#    matches the one in the CM file.
-verify_cm_built_from_seed($config, "CM", "SEED");
+# A. verify a --mapali will work, this will make sure the checksum of the input alignment 
+#    matches the one in the CM file, and save the output alignment
+#    as SEED.1, we'll use this to map RF annotation to the original
+#    SEED.
+verify_cm_built_from_seed($config, "CM", "SEED", "SEED.1");
 
-# 2. verify the COM cmbuild line matches the passed in $desc_opts
+# B. verify the COM cmbuild line matches the passed in $desc_opts
 #    which were read from the DESC file.
 my $desc = $famObj->DESC;
 my $desc_buildopts = $desc->{'BM'};
@@ -148,50 +161,47 @@ $cm_buildopts =~ s/^.+cmbuild/cmbuild/;
 
 if($desc_buildopts ne $cm_buildopts) { die "ERROR cmbuild option/command read from DESC differs from that in CM file: $desc_buildopts != $cm_buildopts"; }
 
-####################################################################################################
-# 3. Uses 'cmbuild' to create a new SEED, with -O option.
+###############################################################
+# 3. Adds RF annotation from new SEED to original SEED to create
+#    a new SEED file, and makes all inserts lowercase and 
+#    consensus residues uppercase, to make even newer SEED.
+#
+# OLD WAY: use cmbuild -O output alignment, this worked on nearly all models but was NOT
+# guaranteed to work on zero basepair models because cmbuild 'doctor's those parsetrees
+# a la HMMER and may shift an insert into a match which eliminates the guarantee that 
+# we can map the RF onto the original SEED. 
+# add_rf_and_ss_cons_given_cmbuild_O("SEED", "SEED.1", "SEED.2", $orig_cm->{cmHeader}->{clen});
+
+# NEW WAY: use cmalign --mapali output alignment, which preserves the input alignment
+# (the original seed in this case) exactly (it never doctors parsetrees (yes, HMMER
+# works the same way)).
+if(-e "SEED") { copy("SEED", "SEED.0"); }
+add_rf_and_ss_cons_given_cmalign_mapali_output("SEED", "SEED.1", "SEED.2", $orig_cm->{cmHeader}->{clen});
+
+##############################################################
+# 4. Builds even newer CM from even newer SEED and verifies 
+#    even newer CM is same as new CM (and also same as original CM).
+my $outfile = "b2.$$.out";
 my $buildopts = $desc_buildopts; 
 $buildopts =~ s/^cmbuild\s+//; # remove 'cmbuild ', cmbuild_wrapper will automatically add this
 $buildopts =~ s/\s*-F\s*/ /;   # remove -F,         cmbuild_wrapper will automatically add this
 $buildopts =~ s/\s*CM\s*/ /;   # remove 'CM',       cmbuild_wrapper will automatically add this
 $buildopts =~ s/\s*SEED\s*/ /; # remove 'SEED',     cmbuild_wrapper will automatically add this
-
-$buildopts .= "-O SEED.1";
-
-if(-e "SEED") { copy("SEED", "SEED.0"); }
 if(-e "CM")   { copy("CM"  , "CM.0"); }
 
-my $outfile = "b.$$.out";
-Bio::Rfam::Infernal::cmbuild_wrapper($config, "$buildopts", "CM.1", "SEED.0", $outfile);
-unlink $outfile;
-
-###################################################################################################
-# 4. Verifies new CM is identical to original CM.
-verify_two_cms_have_identical_parameters("CM.0", "CM.1", 0.0001);
-
-###############################################################
-# 5. Adds RF annotation from new SEED to original SEED to create
-#    a new SEED file, and makes all inserts lowercase and 
-#    consensus residues uppercase, to make even newer SEED.
-add_rf_and_ss_cons_given_cmbuild_O("SEED", "SEED.1", "SEED.2", $orig_cm->{cmHeader}->{clen});
-
-##############################################################
-# 6. Builds even newer CM from even newer SEED and verifies 
-#    even newer CM is same as new CM (and also same as original CM).
-$outfile = "b2.$$.out";
-Bio::Rfam::Infernal::cmbuild_wrapper($config, "$buildopts", "CM.2", "SEED.2", $outfile);
-verify_two_cms_have_identical_parameters("CM.2", "CM.1", 0.00000001);
+Bio::Rfam::Infernal::cmbuild_wrapper($config, "$buildopts", "CM.1", "SEED.2", $outfile);
+verify_two_cms_have_identical_parameters("CM.1", "CM", 0.00000001);
 unlink $outfile;
 
 
 ##############################################################
-# 7. Adds E-value stats from original CM to even newer CM.
-add_evalue_and_com_lines_to_cm("CM.2", "CM.3", \@evalueA, \@comA);
+# 5. Adds E-value stats from original CM to even newer CM.
+add_evalue_and_com_lines_to_cm("CM.1", "CM.2", \@evalueA, \@comA);
 
 ################################################################
-# 8. Validates that we can use --mapali with new CM and new SEED
-if(-e "CM.3") { copy("CM.3", "CM"); }
-else          { die "ERROR, CM.3 does not exist"; }
+# 6. Validates that we can use --mapali with new CM and new SEED
+if(-e "CM.2") { copy("CM.2", "CM"); }
+else          { die "ERROR, CM.2 does not exist"; }
 
 if(-e "SEED.2") { copy("SEED.2", "SEED"); }
 else          { die "ERROR, SEED.2 does not exist"; }
@@ -200,7 +210,7 @@ verify_cm_built_from_seed($config, "CM", "SEED");
 # now CM and SEED are the updated versions
 
 ##############################################################
-# 9. Uses 'touch' to "update" 'TBLOUT', 'SCORES', 'REVTBLOUT', 
+# 7. Uses 'touch' to "update" 'TBLOUT', 'SCORES', 'REVTBLOUT', 
 #    'outlist', 'species', 'taxinfo', 'revspecies', 'outlist.pdf',
 #    etc.
 
@@ -211,8 +221,15 @@ foreach $file (@reqdFiles_t1, @reqdFiles_t2, @reqdFiles_t3) {
   sleep(0.1);
 }
 
+foreach $file (@nonreqdFiles) { 
+  if(-e $file) { 
+    system("touch $file");
+  }
+  sleep(0.1);
+}
+
 # clean up
-foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2", "CM.3") { 
+foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2") { 
   if(-e $file) { unlink $file; }
 }
 
@@ -220,6 +237,105 @@ foreach $file ("SEED.1", "SEED.2", "CM.1", "CM.2", "CM.3") {
 # SUBROUTINES #
 ###############
 
+# add_rf_and_ss_cons_given_cmalign_mapali_output():
+# Given an original SEED alignment ($orig_infile) used to build a CM, and an alignment
+# output from cmalign --mapali ($cmalign_mapali_infile) with a single extra sequence other
+# than the original SEED, add RF and SS_cons annotation to the original SEED alignment 
+# and save it as $outfile.
+sub add_rf_and_ss_cons_given_cmalign_mapali_output {
+
+  my ($orig_infile, $cmalign_mapali_infile, $outfile, $clen) = @_;
+
+  # read in original seed
+  my $orig_seed = Bio::Easel::MSA->new({
+    fileLocation => $orig_infile,
+    forceText    => 1
+      });
+  
+  # read in new seed plus consensus sequence aligned (created with cmalign --mapali SEED CM c.fa)
+  my $tmp_seed  = Bio::Easel::MSA->new({
+    fileLocation => $cmalign_mapali_infile,
+    forceText    => 1
+   });
+  
+  # remove final sequence, the consensus sequence, to get a new alignment
+  my @usemeA = ();
+  my $i;
+  for($i = 0;                  $i < $orig_seed->nseq; $i++) { $usemeA[$i] = 1; }
+  for($i = $orig_seed->nseq(); $i < $tmp_seed->nseq;  $i++) { $usemeA[$i] = 0; }
+  my $new_seed = $tmp_seed->sequence_subset(\@usemeA);
+  undef $tmp_seed;
+  
+  # verify sequences are in the same order in each MSA
+  for(my $i = 0; $i < $orig_seed->nseq(); $i++) { 
+    if($orig_seed->get_sqname($i) ne $new_seed->get_sqname($i)) { 
+      die "ERROR sequences in SEED and SEED.1 are in a different order.";
+    }
+  }
+
+  # Map the two alignments: for each nongap RF column in the $new_seed,
+  # we should be able to unambiguously match an identical column in the $orig_seed.
+  # We'll use this map to create the RF and SS_cons strings for the original
+  # alignments.
+  my $orig_alen = $orig_seed->alen();
+  my $new_alen  = $new_seed->alen();
+  
+  my $new_rf       = $new_seed->get_rf();
+  my @new_rfA      = split("", $new_rf);
+  
+  my $new_ss_cons  = $new_seed->get_ss_cons();
+  my @new_ss_consA = split("", $new_ss_cons);
+  
+  my $cpos         = 0;
+  my $new_apos     = 0;
+  my $orig_apos    = 0;
+  my $orig_rf      = "";
+  my $orig_ss_cons = "";
+  while(($new_apos < $new_alen) && ($cpos < $clen)) { 
+    while($new_rfA[$new_apos] !~ m/[a-zA-Z]/) { 
+      $new_apos++;
+    }
+    $cpos++;
+    #printf("cpos: $cpos new_apos: $new_apos orig_apos: $orig_apos\n");
+    # get this column of the original seed
+    my $new_col = $new_seed->get_column($new_apos+1);
+    # find identical column in new seed
+    my $found_match = 0;
+    while((! $found_match) && ($orig_apos < $orig_alen)) { 
+      my $orig_col = $orig_seed->get_column($orig_apos+1);
+      $orig_col =~ tr/a-z/A-Z/;     # translate to upper case
+      $orig_col =~ s/[^A-Za-z]/-/g; # translate non-alphacharacters to Infernal consensus gap char: '-'
+      if($orig_col eq $new_col) { 
+        $found_match = 1; 
+        # printf("\tfound match new:orig $new_apos:$orig_apos\n"); 
+        $orig_rf      .= "$new_rfA[$new_apos]";
+        $orig_ss_cons .= "$new_ss_consA[$new_apos]";
+      }
+      else { 
+        $orig_rf      .= ".";
+        $orig_ss_cons .= ".";
+      }
+      $orig_apos++;
+    }
+    $new_apos++;
+    if((! $found_match) && ($orig_apos == $orig_alen)) { die "ERROR unable to find match for consensus position $cpos"; }
+  }
+  # we've reached clen, deal with possibility of inserts after final cpos
+  while(length($orig_rf) < $orig_alen) { $orig_rf .= "."; $orig_ss_cons .= "."; }
+
+  $orig_seed->set_rf($orig_rf);
+  $orig_seed->set_ss_cons($orig_ss_cons);
+  $orig_seed->capitalize_based_on_rf();
+  
+  $orig_seed->write_msa($outfile);
+  
+  return;
+}
+
+#######################################################################
+# OLD WAY: we used to map RF and SS_cons using this method but it
+# can fail on zero basepair models because cmbuild doctors parsetrees.
+#
 # add_rf_and_ss_cons_given_cmbuild_O():
 # Given an original SEED alignment ($orig_infile) used to build a CM, and the output -O 
 # alignment from cmbuild ($cmbuild_O_infile) created when the CM was built from the SEED,
@@ -272,7 +388,7 @@ sub add_rf_and_ss_cons_given_cmbuild_O {
       $new_apos++;
     }
     $cpos++;
-    # printf("cpos: $cpos new_apos: $new_apos orig_apos: $orig_apos\n");
+    printf("cpos: $cpos new_apos: $new_apos orig_apos: $orig_apos\n");
     # get this column of the original seed
     my $new_col = $new_seed->get_column($new_apos+1);
     # find identical column in new seed
@@ -433,9 +549,13 @@ sub add_evalue_and_com_lines_to_cm {
 # the CM, with verifying that a --mapali will work, this will make 
 # sure the checksum of the input alignment matches the one in the CM file.
 #
+# If $stkfile is undefined, unlink it at end of function, otherwise
+# save output alignment as $stkfile and don't unlink it.
 sub verify_cm_built_from_seed { 
 
-  my ($config, $cmfile, $seedfile) = @_;
+  my ($config, $cmfile, $seedfile, $stkfile) = @_;
+
+  my $unlink_stkfile;
 
   # 1. verify a --mapali will work, this will make sure the checksum of the input alignment 
   #    matches the one in the CM file.
@@ -445,12 +565,20 @@ sub verify_cm_built_from_seed {
   Bio::Rfam::Infernal::cmemit_wrapper($config, "-c -o $fafile", $cmfile, undef, 0);
 
   # align to CM with --mapali
-  my $stkfile = "$$.stk";
+  if(! defined $stkfile) { 
+    $stkfile = "$$.stk";
+    $unlink_stkfile = 1;
+  }
+  else { 
+    $unlink_stkfile = 0;
+  }
   Bio::Rfam::Infernal::cmalign_wrapper($config, "", "", "--mapali $seedfile -o $stkfile", "CM", $fafile, undef, "", 1, 100, 1, 0, "", 1, undef, 0);
   # we'll die if $seedfile wasn't used to build $cmfile
 
   unlink $fafile;
-  unlink $stkfile;
+  if($unlink_stkfile) { 
+    unlink $stkfile;
+  }
 
   return;
 }
